move moduleNameAsIdentifier to annotator_host
This function is called by all the different parts of JSDoc annotation.
I could maybe stuff it into type_translator to break the circularity
but semantically it's effectively a method on the AnnotatorHost, so
let's put it alongside there.
